Assume k8s:// inCluster kubernetes if  configStore2URL is empty (#1270)
Automatic merge from submit-queue

Assume k8s:// inCluster kubernetes if  configStore2URL is empty

**Release note**:

```release-note
configStore2URL is the URL to connect to the config store. Use k8s://path_to_kubeconfig or fs:// for file system. If path_to_kubeconfig is empty or the parameter is unspecified, in-cluster kubeconfig is used.
```